scripts/requirements.txt
pdfplumber==0.9.0
python-docx==0.8.11
phonenumbers==8.13.18
